#21. Rework Memory service tests.
