+++ Test/Service/Memory/Memory/Test/OrderedPool.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace Emergence::Memory::Test
+{
+bool OrderedPoolTestIncludeMarker () noexcept;
+} // namespace Emergence::Memory::Test
